Changed 2000_CMIP6_AMIP_1deg_ensemble testmods to coldstart
Answers didn't change. Reason for going with coldstart:
The corresponding tests were aborting with variable not found in
initInterpMultilevelContainer.F90 lines 106-114, failing to find
coord_varname = 'COL_Z' and/or level_class_varname = 'LEVGRND_CLASS'.
Instead of reintroducing code that ultimately is unnecessary to SLIM
just to get this function to work, I decided to use coldstart for these
tests.