Add f09 surdat & tests; rm 1deg from testmod names; use if sec == 0
The f09 tests are failing with
ERROR: LND_NTHRDS must be 1 as threading isn't implemented (see issue #14)
so I need to find where cases pick up LND_NTHRDS or how I can override that.

The "if sec == 0" change does not change answers, so I closed #70 where
we were going to figure this out.